+++ src/main/java/pe/edu/upc/medih/servicesinterfaces/IUsuarioService.java
@@ -1,5 +1,7 @@
 package pe.edu.upc.medih.servicesinterfaces;
 
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import pe.edu.upc.medih.entities.Usuario;
 
 import java.util.List;
@@ -8,8 +10,9 @@ public interface IUsuarioService {
     public List<Usuario> list();
     public void insert(Usuario a);
 
-    public Usuario searchbyId(int id);
+    public Usuario searchbyId(long id);
     public void update(Usuario a);
 
-    public void delete(int id);
+    public void delete(long id);
+
 }
